Fixed a few issues
