+++ src/index.js
@@ -1,7 +1,7 @@
+export { VariantChatState } from './types/VariantChatState';
 export { VariantChat } from './components/VariantChat';
 export { useVariantChat } from './lib/VariantChat';
 export {
   initialVariantChatState,
   variantChatReducer,
-  VariantChatState,
 } from './store/slices/chat/chat';
