+++ src/lib/VariantChat.ts
@@ -1,7 +1,7 @@
 import { useApolloClient } from '../hooks/useApolloClient';
 import {
-  doFreshchatInit,
   useFreshchatGetNewMessages,
+  useFreshchatInit,
 } from '../hooks/useFreshchat';
 import { VariantChatConfig } from '../types/VariantChat';
 import { subscribe } from './Event';
@@ -34,7 +34,7 @@ export const useVariantChat = (
   ////////////////////////////////////////////////////
 
   //useFreshchatInit(driverId, config.chatProvider, dispatch);
-  doFreshchatInit(driverId, config.chatProvider, dispatch);
+  useFreshchatInit(driverId, config.chatProvider, dispatch);
 
   useFreshchatGetNewMessages();
 };
