+++ src/store/initialStoreState.ts
@@ -1,4 +1,6 @@
-import { VariantChatState, initialVariantChatState } from './slices/chat/chat';
+import { VariantChatState } from 'types/VariantChatState';
+
+import { initialVariantChatState } from './slices/chat/chat';
 
 export interface StoreState {
   chat: VariantChatState;
