+++ src/store/rootReducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from '@reduxjs/toolkit';
 
-import { chatReducer } from './slices/chat/chat';
+import { variantChatReducer } from './slices/chat/chat';
 
 export const rootReducer = combineReducers({
-  chat: chatReducer,
+  chat: variantChatReducer,
 });
